# openapi.json
{"openapi":"3.1.0","info":{"title":"Tender Intelligence Agent API","description":"Production-grade AI pipeline for public Bus Operations tender analysis, screening, and RAG Q&A","version":"1.0.0"},"paths":{"/api/tenders":{"get":{"tags":["Tenders"],"summary":"List Tenders","operationId":"list_tenders_api_tenders_get","parameters":[{"name":"state","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter tenders by state","title":"State"},"description":"Filter tenders by state"},{"name":"verdict","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter tenders by GO / NO-GO / REVIEW verdict","title":"Verdict"},"description":"Filter tenders by GO / NO-GO / REVIEW verdict"},{"name":"search","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Search term in title or authority","title":"Search"},"description":"Search term in title or authority"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TenderListResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/tenders/daily-digest":{"get":{"tags":["Tenders"],"summary":"Get Daily Digest","operationId":"get_daily_digest_api_tenders_daily_digest_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/tenders/{tender_id}":{"get":{"tags":["Tenders"],"summary":"Get Tender By Id","operationId":"get_tender_by_id_api_tenders__tender_id__get","parameters":[{"name":"tender_id","in":"path","required":true,"schema":{"type":"string","title":"Tender Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TenderResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/tenders/{tender_id}/screening":{"get":{"tags":["Screening"],"summary":"Get Tender Screening","operationId":"get_tender_screening_api_tenders__tender_id__screening_get","parameters":[{"name":"tender_id","in":"path","required":true,"schema":{"type":"string","title":"Tender Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScreeningResultSchema"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/tenders/{tender_id}/screen":{"post":{"tags":["Screening"],"summary":"Run Tender Screening","operationId":"run_tender_screening_api_tenders__tender_id__screen_post","parameters":[{"name":"tender_id","in":"path","required":true,"schema":{"type":"string","title":"Tender Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScreeningResultSchema"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/profile":{"get":{"tags":["Company Profile"],"summary":"Get Company Profile","operationId":"get_company_profile_api_profile_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CompanyProfileResponse"}}}}}},"put":{"tags":["Company Profile"],"summary":"Update Company Profile","operationId":"update_company_profile_api_profile_put","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CompanyProfileUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CompanyProfileResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/chat":{"post":{"tags":["RAG Chat"],"summary":"Chat Qna","operationId":"chat_qna_api_chat_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/ingestion/run":{"post":{"tags":["Ingestion"],"summary":"Trigger Ingestion Run","description":"Triggers asynchronous background ingestion of seed PDF tender documents.\nProtects against concurrent job execution by raising HTTP 409 Conflict if a job is RUNNING.","operationId":"trigger_ingestion_run_api_ingestion_run_post","responses":{"202":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IngestionJobResponse"}}}}}}},"/api/ingestion/{job_id}":{"get":{"tags":["Ingestion"],"summary":"Get Ingestion Job Status","operationId":"get_ingestion_job_status_api_ingestion__job_id__get","parameters":[{"name":"job_id","in":"path","required":true,"schema":{"type":"string","title":"Job Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IngestionJobResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/":{"get":{"tags":["Root"],"summary":"Root Index","description":"Root status endpoint.","operationId":"root_index_api__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api":{"get":{"tags":["Root"],"summary":"Root Index","description":"Root status endpoint.","operationId":"root_index_api_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/":{"get":{"tags":["Root"],"summary":"Root Index","description":"Root status endpoint.","operationId":"root_index__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/health":{"get":{"tags":["Health"],"summary":"Health Check","description":"Basic health check endpoint.","operationId":"health_check_api_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"tags":["Health"],"summary":"Health Check","description":"Basic health check endpoint.","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/tenders":{"get":{"tags":["Tenders Alias"],"summary":"List Tenders Alias","operationId":"list_tenders_alias_tenders_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/ask":{"post":{"tags":["RAG Alias"],"summary":"Ask Alias","operationId":"ask_alias_ask_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/search":{"post":{"tags":["Search Alias"],"summary":"Search Alias","operationId":"search_alias_search_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/ingest":{"post":{"tags":["Ingestion Alias"],"summary":"Ingest Alias","operationId":"ingest_alias_ingest_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/debug-routing":{"get":{"tags":["Debug"],"summary":"Debug Routing","description":"Debug endpoint to inspect exact ASGI path and headers received from Vercel.","operationId":"debug_routing_api_debug_routing_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"ChatRequest":{"properties":{"question":{"type":"string","title":"Question","description":"Natural language question about stored tenders."},"tender_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Tender Id","description":"Optional tender ID filter to constrain Q&A context."}},"type":"object","required":["question"],"title":"ChatRequest"},"ChatResponse":{"properties":{"question":{"type":"string","title":"Question"},"answer":{"type":"string","title":"Answer"},"citations":{"items":{"$ref":"#/components/schemas/Citation"},"type":"array","title":"Citations"},"model_used":{"type":"string","title":"Model Used"},"usage":{"additionalProperties":true,"type":"object","title":"Usage"}},"type":"object","required":["question","answer","citations","model_used","usage"],"title":"ChatResponse"},"Citation":{"properties":{"tender_id":{"type":"string","title":"Tender Id","description":"UUID of the tender source."},"tender_title":{"type":"string","title":"Tender Title","description":"Title of the tender source."},"document_name":{"type":"string","title":"Document Name","description":"File name of the document."},"page_number":{"type":"integer","title":"Page Number","description":"Page number where evidence is found."},"chunk_index":{"type":"integer","title":"Chunk Index","description":"Chunk index number."},"snippet":{"type":"string","title":"Snippet","description":"Verbatim excerpt from the document page."}},"type":"object","required":["tender_id","tender_title","document_name","page_number","chunk_index","snippet"],"title":"Citation"},"CompanyProfileResponse":{"properties":{"fleet_size":{"type":"integer","title":"Fleet Size","description":"Total active operating commercial buses owned/operated.","default":120},"annual_turnover":{"type":"number","title":"Annual Turnover","description":"Annual financial turnover in INR.","default":150000000.0},"years_experience":{"type":"integer","title":"Years Experience","description":"Total years of operating experience in transit.","default":7},"past_contract_sizes":{"items":{"type":"number"},"type":"array","title":"Past Contract Sizes","description":"List of single executed past contract values in INR."},"preferred_geographies":{"items":{"type":"string"},"type":"array","title":"Preferred Geographies","description":"Preferred operational states/cities."},"id":{"type":"string","title":"Id"},"updated_at":{"type":"string","title":"Updated At"}},"type":"object","required":["id","updated_at"],"title":"CompanyProfileResponse"},"CompanyProfileUpdate":{"properties":{"fleet_size":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Fleet Size"},"annual_turnover":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Annual Turnover"},"years_experience":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Years Experience"},"past_contract_sizes":{"anyOf":[{"items":{"type":"number"},"type":"array"},{"type":"null"}],"title":"Past Contract Sizes"},"preferred_geographies":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Preferred Geographies"}},"type":"object","title":"CompanyProfileUpdate"},"CriterionDetail":{"properties":{"criterion_name":{"type":"string","title":"Criterion Name","description":"Name of the evaluated eligibility criterion."},"is_mandatory":{"type":"boolean","title":"Is Mandatory","description":"Whether this criterion is mandatory (hard failure triggers NO-GO)."},"verdict":{"$ref":"#/components/schemas/CriterionVerdict","description":"Individual criterion evaluation verdict: PASS, FAIL, or REVIEW."},"company_value":{"title":"Company Value","description":"Value from the company profile."},"required_value":{"title":"Required Value","description":"Required threshold or clause from the tender."},"reason":{"type":"string","title":"Reason","description":"Detailed explanation of evaluation logic."}},"type":"object","required":["criterion_name","is_mandatory","verdict","company_value","required_value","reason"],"title":"CriterionDetail"},"CriterionVerdict":{"type":"string","enum":["PASS","FAIL","REVIEW"],"title":"CriterionVerdict"},"FinalVerdict":{"type":"string","enum":["GO","NO-GO","REVIEW"],"title":"FinalVerdict"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"IngestionJobResponse":{"properties":{"job_id":{"type":"string","title":"Job Id"},"status":{"type":"string","title":"Status"},"total_documents":{"type":"integer","title":"Total Documents"},"completed_documents":{"type":"integer","title":"Completed Documents"},"failed_documents":{"type":"integer","title":"Failed Documents"},"current_document":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Current Document"},"started_at":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Started At"},"completed_at":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Completed At"},"error_message":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error Message"}},"type":"object","required":["job_id","status","total_documents","completed_documents","failed_documents"],"title":"IngestionJobResponse"},"ScreeningResultSchema":{"properties":{"tender_id":{"type":"string","title":"Tender Id"},"verdict":{"$ref":"#/components/schemas/FinalVerdict"},"reasoning":{"type":"string","title":"Reasoning"},"criteria_results":{"items":{"$ref":"#/components/schemas/CriterionDetail"},"type":"array","title":"Criteria Results"},"screened_at":{"type":"string","title":"Screened At"}},"type":"object","required":["tender_id","verdict","reasoning","criteria_results","screened_at"],"title":"ScreeningResultSchema"},"ScreeningSummaryResponse":{"properties":{"verdict":{"type":"string","title":"Verdict"},"reasoning":{"type":"string","title":"Reasoning"},"criteria_results":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Criteria Results"},"screened_at":{"type":"string","title":"Screened At"}},"type":"object","required":["verdict","reasoning","criteria_results","screened_at"],"title":"ScreeningSummaryResponse"},"TenderEligibilityResponse":{"properties":{"minimum_fleet_size":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Minimum Fleet Size"},"minimum_annual_turnover":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Minimum Annual Turnover"},"minimum_experience_years":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Minimum Experience Years"},"minimum_past_contract_value":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Minimum Past Contract Value"},"minimum_depots_required":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Minimum Depots Required"},"required_geographies":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Required Geographies"},"other_requirements":{"anyOf":[{"items":{"additionalProperties":true,"type":"object"},"type":"array"},{"type":"null"}],"title":"Other Requirements"}},"type":"object","title":"TenderEligibilityResponse"},"TenderListResponse":{"properties":{"total":{"type":"integer","title":"Total"},"tenders":{"items":{"$ref":"#/components/schemas/TenderResponse"},"type":"array","title":"Tenders"}},"type":"object","required":["total","tenders"],"title":"TenderListResponse"},"TenderResponse":{"properties":{"id":{"type":"string","title":"Id"},"title":{"type":"string","title":"Title"},"original_bus_quantity":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Original Bus Quantity"},"latest_bus_quantity":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Latest Bus Quantity"},"latest_quantity_source":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Latest Quantity Source"},"issuing_authority":{"type":"string","title":"Issuing Authority"},"city":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"City"},"state":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"State"},"category":{"type":"string","title":"Category"},"submission_deadline":{"type":"string","title":"Submission Deadline"},"original_deadline":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Original Deadline"},"latest_deadline":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Latest Deadline"},"latest_deadline_source":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Latest Deadline Source"},"timezone":{"type":"string","title":"Timezone"},"days_remaining":{"type":"integer","title":"Days Remaining"},"is_expired":{"type":"boolean","title":"Is Expired"},"emd_amount":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Emd Amount"},"original_emd_amount":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Original Emd Amount"},"latest_emd_amount":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Latest Emd Amount"},"latest_emd_source":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Latest Emd Source"},"emd_breakdown":{"anyOf":[{"additionalProperties":{"type":"number"},"type":"object"},{"type":"null"}],"title":"Emd Breakdown"},"document_fee":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Document Fee"},"scope_summary":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Scope Summary"},"source_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Source Url"},"source_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Source Name"},"document_hash":{"type":"string","title":"Document Hash"},"created_at":{"type":"string","title":"Created At"},"screening":{"anyOf":[{"$ref":"#/components/schemas/ScreeningSummaryResponse"},{"type":"null"}]},"eligibility":{"anyOf":[{"$ref":"#/components/schemas/TenderEligibilityResponse"},{"type":"null"}]}},"type":"object","required":["id","title","issuing_authority","category","submission_deadline","timezone","days_remaining","is_expired","document_hash","created_at"],"title":"TenderResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}